day_15_agents/requirements.txt
langchain
langchain-mistralai
langchain-community
langchain-openai
langchain-ollama
langchainhub
python-dotenv
langchain-mistralai
beautifulsoup4